Extract plugin context to separate file
